Move test to isolated test class to avoid side effects during tests
